+++ google/src/main/scala/org/broadinstitute/dsde/workbench/google/GoogleDirectoryDAO.scala
@@ -2,6 +2,7 @@ package org.broadinstitute.dsde.workbench.google
 
 import com.google.api.services.admin.directory.model.Group
 import org.broadinstitute.dsde.workbench.model._
+import com.google.api.services.groupssettings.model.{Groups => GroupSettings}
 
 import scala.concurrent.Future
 
@@ -13,11 +14,33 @@ trait GoogleDirectoryDAO {
 
   @deprecated(message = "use createGroup(String, WorkbenchEmail) instead", since = "0.9")
   def createGroup(groupName: WorkbenchGroupName, groupEmail: WorkbenchEmail): Future[Unit]
-  def createGroup(displayName: String, groupEmail: WorkbenchEmail): Future[Unit]
+  def createGroup(displayName: String, groupEmail: WorkbenchEmail, groupSettings: Option[GroupSettings] = None): Future[Unit]
   def deleteGroup(groupEmail: WorkbenchEmail): Future[Unit]
   def addMemberToGroup(groupEmail: WorkbenchEmail, memberEmail: WorkbenchEmail): Future[Unit]
   def removeMemberFromGroup(groupEmail: WorkbenchEmail, memberEmail: WorkbenchEmail): Future[Unit]
   def getGoogleGroup(groupEmail: WorkbenchEmail): Future[Option[Group]]
   def isGroupMember(groupEmail: WorkbenchEmail, memberEmail: WorkbenchEmail): Future[Boolean]
   def listGroupMembers(groupEmail: WorkbenchEmail): Future[Option[Seq[String]]]
+  
+  def lockedDownGroupSettings = new GroupSettings()
+    .setWhoCanAdd("ALL_OWNERS_CAN_ADD")
+    .setWhoCanJoin("INVITED_CAN_JOIN")
+    .setWhoCanViewMembership("ALL_MANAGERS_CAN_VIEW")
+    .setWhoCanViewGroup("ALL_OWNERS_CAN_VIEW")
+    .setWhoCanInvite("NONE_CAN_INVITE")
+    .setArchiveOnly("true") //.setWhoCanPostMessage("NONE_CAN_POST")
+    .setWhoCanLeaveGroup("NONE_CAN_LEAVE")
+    .setWhoCanContactOwner("ALL_MANAGERS_CAN_CONTACT")
+    .setWhoCanAddReferences("NONE")
+    .setWhoCanAssignTopics("NONE")
+    .setWhoCanUnassignTopic("NONE")
+    .setWhoCanTakeTopics("NONE")
+    .setWhoCanMarkDuplicate("NONE")
+    .setWhoCanMarkNoResponseNeeded("NONE")
+    .setWhoCanMarkFavoriteReplyOnAnyTopic("NONE")
+    .setWhoCanMarkFavoriteReplyOnOwnTopic("NONE")
+    .setWhoCanUnmarkFavoriteReplyOnAnyTopic("NONE")
+    .setWhoCanEnterFreeFormTags("NONE")
+    .setWhoCanModifyTagsAndCategories("NONE")
+
 }